feat: add memory metrics to gcp
