+++ pkg/types/v1/metrics_test.go
@@ -11,11 +11,11 @@ func TestMetricsParser(t *testing.T) {
 	cpuResource := NewMetric("cpu")
 	assert.NotNil(t, cpuResource)
 
-	cpuResource.SetUsagePercentage(20.54).SetUnitAmount(4.0)
+	cpuResource.SetUsage(20.54).SetUnitAmount(4.0)
 	cpuResource.SetType(CPU).SetResourceUnit(Core)
 	cpuResource.SetEmissions(NewResourceEmission(1056.76, GCO2eqkWh))
 
-	assert.Equal(t, cpuResource.Usage(), Percentage(20.54))
+	assert.Equal(t, cpuResource.Usage(), 20.54)
 	assert.Equal(t, cpuResource.UnitAmount(), float64(4.0))
 	assert.Equal(t, cpuResource.Unit(), Core)
 	assert.Equal(t, cpuResource.Type(), CPU)
